[#1] Refactor: 테이블 리스트 비즈니스 로직 hooks 로 리펙터링
